+++ app/src/main/java/org/linphone/ui/main/recordings/viewmodel/RecordingMediaPlayerViewModel.kt
@@ -175,6 +175,14 @@ class RecordingMediaPlayerViewModel
         }
     }
 
+    @UiThread
+    fun seekTo(position: Int) {
+        coreContext.postOnCoreThread {
+            seekPlaybackTo(position)
+            startPlayback()
+        }
+    }
+
     @WorkerThread
     private fun startPlayback() {
         if (!::player.isInitialized) return
@@ -230,6 +238,18 @@ class RecordingMediaPlayerViewModel
         isPlaying.postValue(false)
         updatePositionJob?.cancel()
         updatePositionJob = null
+    }
+
+    @WorkerThread
+    private fun seekPlaybackTo(position: Int) {
+        if (!::player.isInitialized) return
+
+        if (player.state == Player.State.Closed) {
+            player.open(recordingModel.filePath)
+        }
+
+        Log.i("$TAG Seeking player to position [$position]")
+        player.seek(position)
     }
 
     @WorkerThread